requirements.txt
streamlit
pandas
google-generativeai
rapidfuzz
openpyxl
unidecode
xlsxwriter